backend/requirements.txt
asgiref==3.5.2
Django==2.2.28
djangorestframework==3.13.1
djoser==2.1.0
django-filter==2.4.0
reportlab==3.6.10
Pillow==9.2.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
gunicorn==20.0.4
psycopg2-binary==2.8.6
colorama~=0.4.4